Added user authentication services

+++ Green_Mart_Project/server/routers/cartItem.js
@@ -4,7 +4,8 @@ const db = require('../utils/dbpool');
 const { apiSuccess, apiError } = require('../utils/apiresult');
 
 router.post('/insert', (req, resp) => {
-    const { userId, productId, cartItemQuantity } = req.body;
+    const {productId, cartItemQuantity } = req.body;
+    const userId = req.user.id
     db.query('insert into cartitems(userId,productId,cartItemQuantity)values(?,?,?)',
         [userId, productId, cartItemQuantity],
         (err, result) => {
@@ -26,8 +27,8 @@ router.post('/insert', (req, resp) => {
             }
         })
 })
-router.get("/:userId",(req,resp)=>{
-     const userId=req.params.userId;
+router.get("/",(req,resp)=>{
+     const userId=req.user.id;
      db.query('select * from cartitems where userId=?',[userId],(err,result)=>{
         if(err)
             return resp.send(apiError(err))
@@ -38,7 +39,8 @@ router.get("/:userId",(req,resp)=>{
 router.put("/:cartItemId", (req, resp) => {
     const cartItemQuantity = req.body.cartItemQuantity;
     const cartItemId = req.params.cartItemId;
-    db.query('update cartitems set cartItemQuantity=? where cartItemId=?', [cartItemQuantity, cartItemId], (err, result) => {
+    const userId = req.user.id
+    db.query('update cartitems set cartItemQuantity=? where cartItemId=? and userId = ?', [cartItemQuantity, cartItemId ,userId], (err, result) => {
         if (err)
             return resp.send(apiError(err));
         resp.send(apiSuccess(result));
@@ -47,7 +49,8 @@ router.put("/:cartItemId", (req, resp) => {
 
 router.delete("/:cartItemId", (req, resp) => {
     const cartItemId = req.params.cartItemId;
-    db.query('delete from cartitems where cartItemId=?', [cartItemId], (err, result) => {
+    const userId = req.user.id
+    db.query('delete from cartitems where cartItemId=? and userId = ?', [cartItemId , userId], (err, result) => {
         if (err)
             return resp.send(apiError(err));
         resp.send(apiSuccess(result));
